% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_utils.R
\name{addDropdownOption}
\alias{addDropdownOption}
\title{Add an option to a dropdown widget}
\usage{
addDropdownOption(dropdown, name, p)
}
\arguments{
\item{dropdown}{The dropdown widget}

\item{name}{Name of the option to add}

\item{p}{ggplot to add with this name}
}
\description{
Add an option to a dropdown widget
}
