% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Calculate dimensions for a plot}
\usage{
get_dimensions(width = NULL, height = NULL, aspect.ratio = 16/9)
}
\arguments{
\item{width}{Plot width in pixels.}

\item{height}{Plot height in pixels.}

\item{aspect.ratio}{Plot aspect ratio}
}
\description{
Calculate dimensions for a plot
}
