% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot_to_html}
\alias{ggplot_to_html}
\title{Convert a ggplot object to an html tag}
\usage{
ggplot_to_html(p, width = NULL, height = NULL, img.format = c("svg", "png"))
}
\arguments{
\item{p}{ggplot2 plot}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{img.format}{Image format to generate. Available options: svg, png}
}
\description{
Convert a ggplot object to an html tag
}
