% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot_to_img_png}
\alias{ggplot_to_img_png}
\title{Convert a ggplot2 plot to a base64 encoded png}
\usage{
ggplot_to_img_png(p, width = NULL, height = NULL)
}
\arguments{
\item{p}{ggplot2 plot}

\item{width}{Plot width in pixels.}

\item{height}{Plot height in pixels.}
}
\description{
Convert a ggplot2 plot to a base64 encoded png
}
