% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_utils.R
\name{removeDropdownOption}
\alias{removeDropdownOption}
\title{Remove an option from a dropdown widget}
\usage{
removeDropdownOption(dropdown, name)
}
\arguments{
\item{dropdown}{The dropdown widget}

\item{name}{Name of the option to add}
}
\description{
Remove an option from a dropdown widget
}
