% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_utils.R
\name{addDropdownOption}
\alias{addDropdownOption}
\title{Add an option to a dropdown widget}
\usage{
addDropdownOption(
  dropdown,
  name,
  p,
  width = NULL,
  height = NULL,
  img.format = c("svg", "png")
)
}
\arguments{
\item{dropdown}{The dropdown widget}

\item{name}{Name of the option to add}

\item{p}{ggplot to add with this name}

\item{width}{Plot width in pixels}

\item{height}{Plot height in pixels}

\item{img.format}{Image format to generate. Available options: svg, png}
}
\description{
Add an option to a dropdown widget
}
