% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{dropdown}
\alias{dropdown}
\title{Create an interactive dropdown}
\usage{
dropdown(data = list(), width = NULL, height = NULL)
}
\arguments{
\item{data}{List of options and corresponding ggplot}

\item{width}{Width in pixels (optional, defaults to automatic sizing)}

\item{height}{Height in pixels (optional, defaults to automatic sizing)}
}
\description{
Create an interactive dropdown
}
